tweaked build scripts

+++ build/premake_clean.sh
@@ -1,2 +1,11 @@
-rm -R gmake
-rm -R xcode4
+pushd $(dirname $0) > /dev/null
+
+if [ -d "gmake" ]; then
+    rm -R gmake
+fi
+
+if [ -d "xcode4" ]; then
+  rm -R xcode4
+fi
+
+popd > /dev/null
