+++ script_clean_output.sh
@@ -1,1 +1,7 @@
-rm -R output
+pushd $(dirname $0) > /dev/null
+
+if [ -d "output" ]; then
+    rm -R output
+fi
+
+popd > /dev/null
